feat(trials): refactor trial db queries and types
- Added new functions for retrieving beagle trial details and trials for specific dogs, enhancing the database interaction layer.
- Implemented search functionality for beagle trials with pagination and sorting capabilities.
- Created types for trial search requests and responses to improve type safety and clarity in the codebase.
- Refactored existing trial-related exports to accommodate the new structure and ensure better organization.

Refs: BEJ-91
