+++ packages/db/index.ts
@@ -121,7 +121,6 @@ export {
   type BeagleTrialDetailsResponseDb,
   type BeagleTrialDetailsRowDb,
   type BeagleTrialDogRowDb,
-  type BeagleTrialSearchModeDb,
   type BeagleTrialSearchRequestDb,
   type BeagleTrialSearchResponseDb,
   type BeagleTrialSearchRowDb,
