+++ packages/db/trials/index.ts
@@ -1,17 +1,16 @@
-export {
-  getBeagleTrialDetailsDb,
-  getBeagleTrialsForDogDb,
-  searchBeagleTrialsDb,
-  type BeagleTrialDetailsRequestDb,
-  type BeagleTrialDetailsResponseDb,
-  type BeagleTrialDetailsRowDb,
-  type BeagleTrialDogRowDb,
-  type BeagleTrialSearchModeDb,
-  type BeagleTrialSearchRequestDb,
-  type BeagleTrialSearchResponseDb,
-  type BeagleTrialSearchRowDb,
-  type BeagleTrialSearchSortDb,
-} from "./repository";
+export { getBeagleTrialDetailsDb } from "./get-beagle-trial-details";
+export { getBeagleTrialsForDogDb } from "./get-beagle-trials-for-dog";
+export { searchBeagleTrialsDb } from "./search-beagle-trials";
+export type {
+  BeagleTrialDetailsRequestDb,
+  BeagleTrialDetailsResponseDb,
+  BeagleTrialDetailsRowDb,
+  BeagleTrialDogRowDb,
+  BeagleTrialSearchRequestDb,
+  BeagleTrialSearchResponseDb,
+  BeagleTrialSearchRowDb,
+  BeagleTrialSearchSortDb,
+} from "./types";
 export { getTrialDogPdfDataDb } from "./pdf";
 export {
   upsertKoiratietokantaAjokResultDb,
