+++ packages/db/trials/repository.ts
@@ -1,585 +1,13 @@
-// Prisma repository functions for trial search, event details, and dog-centric
-// trial result lookups used by public trial features and dog profile composition.
-import { DogSex, type Prisma } from "@prisma/client";
-import { BUSINESS_TIME_ZONE, toBusinessDateOnly } from "../core/date-only";
-import { prisma } from "../core/prisma";
-
-export type BeagleTrialSearchSortDb = "date-desc" | "date-asc";
-export type BeagleTrialSearchModeDb = "year" | "range";
-
-export type BeagleTrialSearchRequestDb = {
-  mode: BeagleTrialSearchModeDb;
-  year?: number;
-  dateFrom?: Date;
-  dateTo?: Date;
-  page?: number;
-  pageSize?: number;
-  sort?: BeagleTrialSearchSortDb;
-};
-
-export type BeagleTrialSearchRowDb = {
-  eventDate: Date;
-  eventPlace: string;
-  judge: string | null;
-  dogCount: number;
-};
-
-export type BeagleTrialSearchResponseDb = {
-  mode: BeagleTrialSearchModeDb;
-  year: number | null;
-  dateFrom: Date | null;
-  dateTo: Date | null;
-  availableYears: number[];
-  total: number;
-  totalPages: number;
-  page: number;
-  items: BeagleTrialSearchRowDb[];
-};
-
-export type BeagleTrialDetailsRequestDb = {
-  eventDate: Date;
-  eventPlace: string;
-};
-
-export type BeagleTrialDetailsRowDb = {
-  id: string;
-  dogId: string;
-  registrationNo: string;
-  name: string;
-  sex: "U" | "N" | "-";
-  weather: string | null;
-  award: string | null;
-  classCode: string | null;
-  rank: string | null;
-  points: number | null;
-  judge: string | null;
-  // Legacy per-dog scoring fields from `akoeall` rows.
-  // TAPPA -> eventPlace (event-level), these remain dog-row metrics.
-  haku: number | null;
-  hauk: number | null;
-  yva: number | null;
-  hlo: number | null;
-  alo: number | null;
-  tja: number | null;
-  pin: number | null;
-  // Internal import-trace fields (not exposed in public contract/UI).
-  legacyFlag: string | null;
-  sourceKey: string;
-  createdAt: Date;
-  updatedAt: Date;
-};
-
-export type BeagleTrialDetailsResponseDb = {
-  eventDate: Date;
-  eventPlace: string;
-  judge: string | null;
-  dogCount: number;
-  items: BeagleTrialDetailsRowDb[];
-};
-
-export type BeagleTrialDogRowDb = {
-  id: string;
-  place: string;
-  date: Date;
-  weather: string | null;
-  className: string | null;
-  classCode: string | null;
-  rank: string | null;
-  points: number | null;
-  award: string | null;
-  judge: string | null;
-  haku: number | null;
-  hauk: number | null;
-  yva: number | null;
-  hlo: number | null;
-  alo: number | null;
-  tja: number | null;
-  pin: number | null;
-};
-
-function parseIsoDateOnlyParts(
-  value: string,
-): { year: number; month: number; day: number } | null {
-  const match = /^(\d{4})-(\d{2})-(\d{2})$/.exec(value);
-  if (!match) return null;
-
-  const year = Number.parseInt(match[1], 10);
-  const month = Number.parseInt(match[2], 10);
-  const day = Number.parseInt(match[3], 10);
-  if (
-    !Number.isFinite(year) ||
-    !Number.isFinite(month) ||
-    !Number.isFinite(day)
-  ) {
-    return null;
-  }
-
-  const utcDate = new Date(Date.UTC(year, month - 1, day, 0, 0, 0, 0));
-  if (
-    utcDate.getUTCFullYear() !== year ||
-    utcDate.getUTCMonth() !== month - 1 ||
-    utcDate.getUTCDate() !== day
-  ) {
-    return null;
-  }
-
-  return { year, month, day };
-}
-
-function getTimeZoneOffsetMs(date: Date, timeZone: string): number {
-  const formatter = new Intl.DateTimeFormat("en-US", {
-    timeZone,
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-    second: "2-digit",
-    hourCycle: "h23",
-    hour12: false,
-  });
-  const parts = formatter.formatToParts(date);
-  const year = Number.parseInt(
-    parts.find((part) => part.type === "year")?.value ?? "0",
-    10,
-  );
-  const month = Number.parseInt(
-    parts.find((part) => part.type === "month")?.value ?? "0",
-    10,
-  );
-  const day = Number.parseInt(
-    parts.find((part) => part.type === "day")?.value ?? "0",
-    10,
-  );
-  const hour = Number.parseInt(
-    parts.find((part) => part.type === "hour")?.value ?? "0",
-    10,
-  );
-  const minute = Number.parseInt(
-    parts.find((part) => part.type === "minute")?.value ?? "0",
-    10,
-  );
-  const second = Number.parseInt(
-    parts.find((part) => part.type === "second")?.value ?? "0",
-    10,
-  );
-  const normalizedHour = hour === 24 ? 0 : hour;
-  const asUtc = Date.UTC(
-    year,
-    month - 1,
-    day,
-    normalizedHour,
-    minute,
-    second,
-    0,
-  );
-  return asUtc - date.getTime();
-}
-
-function toBusinessDateStartUtc(isoDate: string): Date | null {
-  const parsed = parseIsoDateOnlyParts(isoDate);
-  if (!parsed) {
-    return null;
-  }
-
-  const { year, month, day } = parsed;
-  const midnightAsUtcMs = Date.UTC(year, month - 1, day, 0, 0, 0, 0);
-  let utcTimeMs = midnightAsUtcMs;
-
-  for (let index = 0; index < 3; index += 1) {
-    const offsetMs = getTimeZoneOffsetMs(
-      new Date(utcTimeMs),
-      BUSINESS_TIME_ZONE,
-    );
-    utcTimeMs = midnightAsUtcMs - offsetMs;
-  }
-
-  return new Date(utcTimeMs);
-}
-
-function addIsoDateDays(isoDate: string, days: number): string | null {
-  const parsed = parseIsoDateOnlyParts(isoDate);
-  if (!parsed) {
-    return null;
-  }
-  const base = new Date(
-    Date.UTC(parsed.year, parsed.month - 1, parsed.day, 0, 0, 0, 0),
-  );
-  base.setUTCDate(base.getUTCDate() + days);
-  return base.toISOString().slice(0, 10);
-}
-
-function normalizeUtcDateToBusinessDateStart(value: Date): Date | null {
-  const isoDate = toBusinessDateOnly(value);
-  return toBusinessDateStartUtc(isoDate);
-}
-
-function parsePage(value: number | undefined): number {
-  if (!Number.isFinite(value)) return 1;
-  return Math.max(1, Math.floor(value ?? 1));
-}
-
-function parsePageSize(value: number | undefined): number {
-  if (!Number.isFinite(value)) return 10;
-  return Math.min(100, Math.max(1, Math.floor(value ?? 10)));
-}
-
-function resolvePagination(
-  total: number,
-  page: number,
-  pageSize: number,
-): {
-  totalPages: number;
-  page: number;
-  start: number;
-} {
-  if (total === 0) {
-    return { totalPages: 0, page: 1, start: 0 };
-  }
-
-  const totalPages = Math.ceil(total / pageSize);
-  const clampedPage = Math.min(Math.max(1, page), totalPages);
-  return {
-    totalPages,
-    page: clampedPage,
-    start: (clampedPage - 1) * pageSize,
-  };
-}
-
-function normalizeSort(
-  value: BeagleTrialSearchSortDb | undefined,
-): BeagleTrialSearchSortDb {
-  return value === "date-asc" ? "date-asc" : "date-desc";
-}
-
-function buildWhere(
-  input: BeagleTrialSearchRequestDb,
-): Prisma.TrialResultWhereInput {
-  if (input.mode === "year" && Number.isFinite(input.year)) {
-    const year = Math.floor(input.year ?? 0);
-    const start = toBusinessDateStartUtc(`${year}-01-01`);
-    const end = toBusinessDateStartUtc(`${year + 1}-01-01`);
-    if (!start || !end) {
-      return {};
-    }
-    return {
-      eventDate: {
-        gte: start,
-        lt: end,
-      },
-    };
-  }
-
-  if (input.mode === "range" && input.dateFrom && input.dateTo) {
-    const start = normalizeUtcDateToBusinessDateStart(input.dateFrom);
-    const end = normalizeUtcDateToBusinessDateStart(input.dateTo);
-    if (!start || !end) {
-      return {};
-    }
-
-    return {
-      eventDate: {
-        gte: start,
-        lt: end,
-      },
-    };
-  }
-
-  return {};
-}
-
-const BUSINESS_YEAR_FORMATTER = new Intl.DateTimeFormat("en-US", {
-  timeZone: BUSINESS_TIME_ZONE,
-  year: "numeric",
-});
-
-function toBusinessYear(value: Date): number {
-  const year = BUSINESS_YEAR_FORMATTER.formatToParts(value).find(
-    (part) => part.type === "year",
-  )?.value;
-  if (!year) return value.getUTCFullYear();
-  const parsed = Number.parseInt(year, 10);
-  return Number.isFinite(parsed) ? parsed : value.getUTCFullYear();
-}
-
-function compareRows(
-  left: BeagleTrialSearchRowDb,
-  right: BeagleTrialSearchRowDb,
-  sort: BeagleTrialSearchSortDb,
-): number {
-  const dateComparison =
-    sort === "date-asc"
-      ? left.eventDate.getTime() - right.eventDate.getTime()
-      : right.eventDate.getTime() - left.eventDate.getTime();
-  if (dateComparison !== 0) return dateComparison;
-  return left.eventPlace.localeCompare(right.eventPlace, "fi", {
-    sensitivity: "base",
-  });
-}
-
-export async function searchBeagleTrialsDb(
-  input: BeagleTrialSearchRequestDb,
-): Promise<BeagleTrialSearchResponseDb> {
-  const page = parsePage(input.page);
-  const pageSize = parsePageSize(input.pageSize);
-  const sort = normalizeSort(input.sort);
-  const where = buildWhere(input);
-
-  const [yearRows, groupedRows] = await Promise.all([
-    prisma.trialResult.groupBy({
-      by: ["eventDate"],
-    }),
-    prisma.trialResult.groupBy({
-      by: ["eventDate", "eventPlace"],
-      where,
-      _count: { _all: true },
-      _max: { judge: true },
-    }),
-  ]);
-
-  const availableYears = Array.from(
-    new Set(yearRows.map((row) => toBusinessYear(row.eventDate))),
-  ).sort((left, right) => right - left);
-
-  const rows: BeagleTrialSearchRowDb[] = groupedRows
-    .map((row) => ({
-      eventDate: row.eventDate,
-      eventPlace: row.eventPlace,
-      judge: row._max.judge ?? null,
-      dogCount: row._count._all,
-    }))
-    .sort((left, right) => compareRows(left, right, sort));
-
-  const total = rows.length;
-  const pagination = resolvePagination(total, page, pageSize);
-  const pageRows = rows.slice(pagination.start, pagination.start + pageSize);
-
-  const judgeRows =
-    pageRows.length === 0
-      ? []
-      : await prisma.trialResult.findMany({
-          where: {
-            OR: pageRows.map((row) => ({
-              eventDate: row.eventDate,
-              eventPlace: row.eventPlace,
-            })),
-          },
-          select: {
-            eventDate: true,
-            eventPlace: true,
-            judge: true,
-          },
-        });
-
-  const judgeByEvent = new Map<string, string | null>();
-  for (const row of judgeRows) {
-    const key = `${row.eventDate.toISOString()}::${row.eventPlace}`;
-    const normalizedJudge = row.judge?.trim() ?? "";
-    if (!normalizedJudge) {
-      continue;
-    }
-    const existing = judgeByEvent.get(key);
-    if (existing == null) {
-      judgeByEvent.set(key, normalizedJudge);
-      continue;
-    }
-    if (existing !== normalizedJudge) {
-      judgeByEvent.set(key, null);
-    }
-  }
-
-  const items = pageRows.map((row) => {
-    const key = `${row.eventDate.toISOString()}::${row.eventPlace}`;
-    const judge = judgeByEvent.get(key);
-    return {
-      ...row,
-      judge: judge === undefined ? null : judge,
-    };
-  });
-
-  return {
-    mode: input.mode,
-    year:
-      input.mode === "year" && Number.isFinite(input.year)
-        ? Math.floor(input.year ?? 0)
-        : null,
-    dateFrom: input.mode === "range" ? (input.dateFrom ?? null) : null,
-    dateTo: input.mode === "range" ? (input.dateTo ?? null) : null,
-    availableYears,
-    total,
-    totalPages: pagination.totalPages,
-    page: pagination.page,
-    items,
-  };
-}
-
-function toSexCode(value: DogSex): "U" | "N" | "-" {
-  if (value === DogSex.MALE) return "U";
-  if (value === DogSex.FEMALE) return "N";
-  return "-";
-}
-
-function compareDetailRows(
-  left: BeagleTrialDetailsRowDb,
-  right: BeagleTrialDetailsRowDb,
-): number {
-  const leftPoints = left.points ?? -1;
-  const rightPoints = right.points ?? -1;
-  if (leftPoints !== rightPoints) {
-    return rightPoints - leftPoints;
-  }
-
-  const rankComparison = (left.rank ?? "").localeCompare(
-    right.rank ?? "",
-    "fi",
-    {
-      sensitivity: "base",
-    },
-  );
-  if (rankComparison !== 0) return rankComparison;
-
-  const nameComparison = left.name.localeCompare(right.name, "fi", {
-    sensitivity: "base",
-  });
-  if (nameComparison !== 0) return nameComparison;
-
-  return left.registrationNo.localeCompare(right.registrationNo, "fi", {
-    sensitivity: "base",
-  });
-}
-
-export async function getBeagleTrialDetailsDb(
-  input: BeagleTrialDetailsRequestDb,
-): Promise<BeagleTrialDetailsResponseDb | null> {
-  const eventDateIso = toBusinessDateOnly(input.eventDate);
-  const rangeStart = toBusinessDateStartUtc(eventDateIso);
-  const nextEventDateIso = addIsoDateDays(eventDateIso, 1);
-  const rangeEnd = nextEventDateIso
-    ? toBusinessDateStartUtc(nextEventDateIso)
-    : null;
-  if (!rangeStart || !rangeEnd) {
-    return null;
-  }
-
-  const rows = await prisma.trialResult.findMany({
-    where: {
-      eventDate: {
-        gte: rangeStart,
-        lt: rangeEnd,
-      },
-      eventPlace: input.eventPlace,
-    },
-    include: {
-      dog: {
-        select: {
-          id: true,
-          name: true,
-          sex: true,
-          registrations: {
-            select: {
-              registrationNo: true,
-            },
-            orderBy: [{ createdAt: "asc" }, { registrationNo: "asc" }],
-            take: 1,
-          },
-        },
-      },
-    },
-  });
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const items: BeagleTrialDetailsRowDb[] = rows
-    .map((row) => ({
-      id: row.id,
-      dogId: row.dog.id,
-      registrationNo: row.dog.registrations[0]?.registrationNo ?? "-",
-      name: row.dog.name,
-      sex: toSexCode(row.dog.sex),
-      // Legacy akoeall result columns shown/copied to users.
-      weather: row.ke,
-      award: row.pa,
-      classCode: row.lk,
-      rank: row.sija,
-      points: row.piste ? row.piste.toNumber() : null,
-      judge: row.judge,
-      haku: row.haku ? row.haku.toNumber() : null,
-      hauk: row.hauk ? row.hauk.toNumber() : null,
-      yva: row.yva ? row.yva.toNumber() : null,
-      hlo: row.hlo ? row.hlo.toNumber() : null,
-      alo: row.alo ? row.alo.toNumber() : null,
-      tja: row.tja ? row.tja.toNumber() : null,
-      pin: row.pin ? row.pin.toNumber() : null,
-      legacyFlag: row.legacyFlag,
-      sourceKey: row.sourceKey,
-      createdAt: row.createdAt,
-      updatedAt: row.updatedAt,
-    }))
-    .sort(compareDetailRows);
-
-  const judges = Array.from(
-    new Set(
-      rows
-        .map((row) => row.judge?.trim())
-        .filter((judge): judge is string => Boolean(judge)),
-    ),
-  );
-
-  return {
-    eventDate: rows[0].eventDate,
-    eventPlace: rows[0].eventPlace,
-    judge: judges.length === 1 ? judges[0] : null,
-    dogCount: items.length,
-    items,
-  };
-}
-
-export async function getBeagleTrialsForDogDb(
-  dogId: string,
-): Promise<BeagleTrialDogRowDb[]> {
-  const rows = await prisma.trialResult.findMany({
-    where: { dogId },
-    select: {
-      id: true,
-      eventPlace: true,
-      eventDate: true,
-      ke: true,
-      eventName: true,
-      lk: true,
-      sija: true,
-      piste: true,
-      pa: true,
-      judge: true,
-      haku: true,
-      hauk: true,
-      yva: true,
-      hlo: true,
-      alo: true,
-      tja: true,
-      pin: true,
-    },
-    orderBy: [{ eventDate: "desc" }, { eventPlace: "asc" }, { id: "asc" }],
-  });
-
-  return rows.map((row) => ({
-    id: row.id,
-    place: row.eventPlace,
-    date: row.eventDate,
-    weather: row.ke,
-    className: row.eventName,
-    classCode: row.lk,
-    rank: row.sija,
-    points: row.piste ? row.piste.toNumber() : null,
-    award: row.pa,
-    judge: row.judge,
-    haku: row.haku ? row.haku.toNumber() : null,
-    hauk: row.hauk ? row.hauk.toNumber() : null,
-    yva: row.yva ? row.yva.toNumber() : null,
-    hlo: row.hlo ? row.hlo.toNumber() : null,
-    alo: row.alo ? row.alo.toNumber() : null,
-    tja: row.tja ? row.tja.toNumber() : null,
-    pin: row.pin ? row.pin.toNumber() : null,
-  }));
-}
+export { getBeagleTrialDetailsDb } from "./get-beagle-trial-details";
+export { getBeagleTrialsForDogDb } from "./get-beagle-trials-for-dog";
+export { searchBeagleTrialsDb } from "./search-beagle-trials";
+export type {
+  BeagleTrialDetailsRequestDb,
+  BeagleTrialDetailsResponseDb,
+  BeagleTrialDetailsRowDb,
+  BeagleTrialDogRowDb,
+  BeagleTrialSearchRequestDb,
+  BeagleTrialSearchResponseDb,
+  BeagleTrialSearchRowDb,
+  BeagleTrialSearchSortDb,
+} from "./types";
