+++ packages/server/trials/get-beagle-trial-details.ts
@@ -4,6 +4,7 @@ import { toBusinessDateOnly } from "../core/date-only";
 import { toErrorLog, withLogContext } from "../core/logger";
 import type { ServiceResult } from "../core/result";
 import { formatTrialAward } from "./core";
+import { getTrialBusinessDateUtcRange } from "./internal/business-date";
 import { encodeTrialId, parseTrialId } from "./internal/trial-id";
 import type { TrialsServiceLogContext } from "./types";
 
@@ -41,8 +42,12 @@ export async function getBeagleTrialDetailsService(
   );
 
   try {
+    const eventDateRange = getTrialBusinessDateUtcRange(
+      parsedTrialId.eventDate,
+    );
     const result = await getBeagleTrialDetailsDb({
-      eventDate: parsedTrialId.eventDate,
+      eventDateStart: eventDateRange.start,
+      eventDateEndExclusive: eventDateRange.endExclusive,
       eventPlace: parsedTrialId.eventPlace,
     });
     if (!result) {
